Changed the CombatStrategy, deleted the player from the constructor, now it is settable via the method setCurrentPlayer. Also added the weapon and move indexes and their getters.
For the ConcreteTurnQueue class we added the reorder method inside the add method.

In the Enemy class we added some prototypes for the new features.

In the Maps class there is now a List of enemies that is in the constructor with his getter method, the turnQueue can be also obtained via the getter.

In the method startMap we added the inizialization of the queue.

Added the checkAggro method for the enemies that needs to be added to the turnQueue.

Used the checkAggro method inside the acceptMovement of the MovementStrategy.

Created the Tests for CombatStrategy and updated the acceptMovement one inside the movementstrategyTest.

Co-authored-by: Davide11000 <Davide11000@users.noreply.github.com>
Co-authored-by: alex-grg <alex-grg@users.noreply.github.com>
Co-authored-by: miriam <mmlst@users.noreply.github.com>